+++ Changelog.md
@@ -27,6 +27,8 @@ OpenCore Changelog
 - Improved ACPI PM timer CPU frequency calculation performance
 - Improved LapicKernelPanic compatibility with newer macOS versions
 - Fixed drivers starting with `#` not being skipped
+- Added audio support through AudioDxe with optional boot chime
+- Added VoiceOver accessability support in boot.efi for 10.13+
 
 #### v0.5.5
 - Fixed CPU bus ratio calculation for Nehalem and Westmere